+++ chainerx_cc/chainerx/cuda/cuda_device/activation.cu
@@ -13,6 +13,7 @@
 #include "chainerx/cuda/op_regist.h"
 #include "chainerx/device.h"
 #include "chainerx/dtype.h"
+#include "chainerx/kernels/math.h"
 #include "chainerx/numeric.h"
 #include "chainerx/routines/math.h"
 #include "chainerx/routines/type_util.h"
