+++ chainerx_cc/chainerx/cuda/cuda_device/dot.cu
@@ -24,6 +24,7 @@
 #include "chainerx/float16.h"
 #include "chainerx/kernels/creation.h"
 #include "chainerx/kernels/linalg.h"
+#include "chainerx/kernels/math.h"
 #include "chainerx/macro.h"
 #include "chainerx/routines/creation.h"
 #include "chainerx/routines/math.h"
