+++ chainerx_cc/chainerx/cuda/cuda_device/misc.cu
@@ -13,7 +13,7 @@
 #include "chainerx/cuda/op_regist.h"
 #include "chainerx/device.h"
 #include "chainerx/dtype.h"
-#include "chainerx/routines/math.h"
+#include "chainerx/kernels/math.h"
 
 namespace chainerx {
 namespace cuda {
