+++ chainerx_cc/chainerx/cuda/cuda_device/trigonometric.cu
@@ -13,8 +13,8 @@
 #include "chainerx/cuda/op_regist.h"
 #include "chainerx/device.h"
 #include "chainerx/dtype.h"
+#include "chainerx/kernels/math.h"
 #include "chainerx/numeric.h"
-#include "chainerx/routines/math.h"
 #include "chainerx/scalar.h"
 
 namespace chainerx {
