+++ chainerx_cc/chainerx/native/native_device/trigonometric.cc
@@ -6,10 +6,10 @@
 #include "chainerx/array.h"
 #include "chainerx/device.h"
 #include "chainerx/dtype.h"
+#include "chainerx/kernels/math.h"
 #include "chainerx/native/elementwise.h"
 #include "chainerx/native/op_regist.h"
 #include "chainerx/numeric.h"
-#include "chainerx/routines/math.h"
 #include "chainerx/scalar.h"
 
 namespace chainerx {
