+++ chainerx_cc/chainerx/routines/indexing.cc
@@ -18,6 +18,7 @@
 #include "chainerx/dtype.h"
 #include "chainerx/graph.h"
 #include "chainerx/kernels/indexing.h"
+#include "chainerx/kernels/math.h"
 #include "chainerx/macro.h"
 #include "chainerx/routines/creation.h"
 #include "chainerx/routines/math.h"
